consultation ticket psy

+++ database/migrations/2020_01_09_105206_update_user_table.php
@@ -14,7 +14,7 @@ class UpdateUserTable extends Migration
     public function up()
     {
         Schema::table('users', function(Blueprint $table){
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('role_id');
 
             $table->foreign('role_id')->references('id')->on('roles');
         });
